set minimal value of seats in bus; delete bus if exists

+++ src/main/java/pl/bussystem/domain/bus/service/BusServiceImpl.java
@@ -25,7 +25,9 @@ public class BusServiceImpl implements BusService {
   @Override
   public void removeByRegistrationNumber(String registrationNumber) {
     BusEntity busEntity = busRepository.findByRegistrationNumber(registrationNumber);
-    busRepository.delete(busEntity);
+    if (busEntity != null) {
+      busRepository.delete(busEntity);
+    }
   }
 
   @Override
